fix(KEN-674): the parser and the copy name [no-changelog]
Address PR comments - two gaps in fixes this branch already made.

A `repo-effects:` line typed without its colon opens no frontmatter
entry: `parse_tolerant` drops the line and the block indented under it,
warns, and moves on. The key is then missing the same way an absent one
is, so the three-state reader called the declaration Absent and a
removal skipped the uninstaller it had named. `Parsed` now carries the
lines it ignored, and a declaration whose key is among them reads as
Unreadable, which stops the plan with the package installed.

`installed_tree` probed every harness's own directory under the shared
tree's spelling. A copy delivery writes `rendered_name`, which joins a
namespaced name with the separator that harness will load - a hyphen
under Copilot and OpenCode, two underscores in the shared tree - so a
namespaced copy read as declaring nothing. Each harness directory is
probed under that harness's own name; the shared tree keeps the
canonical one.

Both files crossed the size threshold with the new tests, so each one's
tests move to the sibling file the rest of the crate keeps them in. No
test changed in the move.
