+++ crates/core/src/frontmatter/tests.rs
@@ -106,3 +106,62 @@ fn scalars_stay_strings_and_null_forms_collapse() {
     assert_eq!(map.get("c"), Some(&Value::Null));
     assert_eq!(map.get("d").and_then(Value::as_str), Some("007"));
 }
+
+fn span_of(text: &str) -> Result<&str, NameProblem> {
+    name_value_span(text).map(|span| &text[span])
+}
+
+#[test]
+fn finds_the_inline_value_and_only_it() {
+    assert_eq!(span_of("---\nname: gh\n---\nBody.\n"), Ok("gh"));
+    assert_eq!(span_of("---\nname : gh\n---\n"), Ok("gh"));
+    assert_eq!(span_of("---\nname:   gh  \n---\n"), Ok("gh"));
+    assert_eq!(span_of("---\r\nname: gh\r\n---\r\n"), Ok("gh"));
+    assert_eq!(span_of("---\nname: gh #edited\n...\n"), Ok("gh #edited"));
+    assert_eq!(span_of("---\nname: \"gh\"\n---\n"), Ok("\"gh\""));
+    assert_eq!(span_of("---\nname: 'gh'\n---\n"), Ok("'gh'"));
+    // A comment after a quoted value belongs to the line, not the
+    // value; an escaped quote does not close the scalar.
+    assert_eq!(span_of("---\nname: \"gh\" # package\n---\n"), Ok("\"gh\""));
+    assert_eq!(span_of("---\nname: 'it''s' # x\n---\n"), Ok("'it''s'"));
+    assert_eq!(
+        span_of("---\nname: \"a\\\"b\" # c\n---\n"),
+        Ok("\"a\\\"b\"")
+    );
+    // A comment-only or blank line after the entry is not a
+    // continuation of its value, indented or not.
+    assert_eq!(span_of("---\nname: gh\n  # note\n---\n"), Ok("gh"));
+    assert_eq!(span_of("---\nname: gh\n\ndesc: d\n---\n"), Ok("gh"));
+    // Not a top-level entry, and not the `name` key.
+    assert_eq!(
+        span_of("---\nmeta:\n  name: inner\nname: outer\n---\n"),
+        Ok("outer")
+    );
+    assert_eq!(
+        span_of("---\nnames: many\n---\n"),
+        Err(NameProblem::Missing { insert_at: 4 })
+    );
+}
+
+#[test]
+fn refuses_what_is_not_one_scalar() {
+    assert_eq!(span_of("Body.\n"), Err(NameProblem::NoFrontmatter));
+    assert_eq!(
+        span_of("---\nname: a\nname: b\n---\n"),
+        Err(NameProblem::Twice)
+    );
+    for text in [
+        "---\nname: [copy]\n---\n",
+        "---\nname: |\n  gh\n---\n",
+        "---\nname: >\n  gh\n---\n",
+        "---\nname: &anchor gh\n---\n",
+        "---\nname: gh\n  continued\n---\n",
+        "---\nname: gh\n  # note\n  continued\n---\n",
+        "---\nname:\n---\n",
+        "---\nname: \"gh\n---\n",
+        "---\nname: \"gh\" trailing\n---\n",
+        "---\nname: \"gh\"#glued\n---\n",
+    ] {
+        assert_eq!(span_of(text), Err(NameProblem::NotAScalar), "{text:?}");
+    }
+}
